pyproject.toml
[project]
name = "bookcatalog"
version = "0.1.0"
description = "AI-powered book cataloging system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "mcp[cli]>=1.26.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[tool.setuptools.packages.find]
include = ["bookcatalog*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
